# src/api/content-api-openapi.yaml
openapi: 3.0.3
info:
  title: MLG.clan Content Storage and Retrieval API
  description: |
    Comprehensive gaming platform API for content storage, retrieval, and management.
    
    ## Features
    - Content CRUD operations for gaming clips, screenshots, guides, and reviews
    - MLG token integration with Solana blockchain voting system
    - Advanced search and filtering capabilities
    - Content moderation and reporting system
    - Real-time analytics and metrics tracking
    - Scalable file upload/download handling
    - User content management and statistics
    
    ## Authentication
    - JWT tokens for user authentication
    - Solana wallet signature verification for blockchain operations
    - Rate limiting and abuse prevention
    
    ## MLG Token Integration
    - Contract Address: 7XJiwLDrjzxDYdZipnJXzpr1iDTmK55XixSFAa7JgNEL
    - Vote costs: Upvote (1 MLG), Downvote (2 MLG), Super Vote (5 MLG)
    - Reward multipliers based on content type and engagement
    
  version: 1.0.0
  contact:
    name: MLG.clan Development Team
    url: https://mlg.clan/api-support
    email: api-support@mlg.clan
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mlg.clan/v1
    description: Production API Server
  - url: https://staging-api.mlg.clan/v1
    description: Staging API Server
  - url: http://localhost:3000/api/v1
    description: Local Development Server
security:
  - BearerAuth: []
  - WalletSignature: []
paths:
...
  /content:
    get:
      summary: List content with filtering and pagination
      description: |
        Retrieve content list with advanced filtering, sorting, and search capabilities.
        Supports both offset-based and cursor-based pagination for optimal performance.
      tags:
        - Content
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - name: contentType
          in: query
          description: Filter by content type
          schema:
            $ref: '#/components/schemas/ContentType'
        - name: game
          in: query
          description: Filter by game title
          schema:
            type: string
            example: "Fortnite"
        - name: platform
          in: query
          description: Filter by gaming platform
          schema:
            $ref: '#/components/schemas/Platform'
        - name: category
          in: query
          description: Filter by content category
          schema:
            $ref: '#/components/schemas/Category'
        - name: userId
          in: query
          description: Filter by content creator
          schema:
            type: string
            format: uuid
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
            example: "fortnite,victory-royale,clutch"
        - name: status
          in: query
          description: Filter by content status
          schema:
            $ref: '#/components/schemas/ContentStatus'
        - name: search
          in: query
          description: Full-text search query
          schema:
            type: string
            maxLength: 100
            example: "epic victory royale"
        - name: searchFields
          in: query
          description: Fields to search in
          schema:
            type: string
            enum: [title, description, tags, title,description, title,tags, description,tags, title,description,tags]
            default: "title,description,tags"
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [created_at, updated_at, views, likes, trending_score, quality_score]
            default: created_at
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: createdAfter
          in: query
          description: Filter content created after this date
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          description: Filter content created before this date
          schema:
            type: string
            format: date-time
        - name: minQualityScore
          in: query
          description: Minimum quality score
          schema:
            type: number
            minimum: 0
            maximum: 100
        - name: minViews
          in: query
          description: Minimum view count
          schema:
            type: integer
            minimum: 0
        - name: minLikes
          in: query
          description: Minimum like count
          schema:
            type: integer
            minimum: 0
        - name: trending
          in: query
          description: Only return trending content
          schema:
            type: boolean
        - name: featured
          in: query
          description: Only return featured content
          schema:
            type: boolean
      responses:
        '200':
          description: Content list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            type: array
                            items:
                              $ref: '#/components/schemas/Content'
                          pagination:
                            $ref: '#/components/schemas/PaginationResponse'
                          filters:
                            type: object
                            properties:
                              applied:
                                type: object
                                description: Currently applied filters
                              available:
                                type: object
                                description: Available filter options
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    
    post:
      summary: Create new content
      description: |
        Upload and create new gaming content with metadata and files.
        Supports chunked upload for large files and provides real-time processing status.
      tags:
        - Content
      security:
        - BearerAuth: []
        - WalletSignature: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - metadata
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
                  description: Main content files (max 5 files, 500MB total)
                thumbnail:
                  type: string
                  format: binary
                  description: Optional custom thumbnail image
                metadata:
                  $ref: '#/components/schemas/ContentCreateInput'
            encoding:
              files:
                contentType: video/*, image/*, application/pdf, text/*
              thumbnail:
                contentType: image/*
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            $ref: '#/components/schemas/Content'
                          uploadUrls:
                            type: array
                            items:
                              type: string
                              format: uri
                            description: CDN upload URLs for additional processing
                          processingStatus:
                            type: string
                            enum: [initiated, processing, complete, failed]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    
    WalletSignature:
      type: apiKey
      in: header
      name: X-Wallet-Signature
      description: |
        Solana wallet signature for blockchain operations.
        Required format: {signature}:{message}:{publicKey}
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    Cursor:
      name: cursor
      in: query
      description: Cursor for pagination
      schema:
        type: string
  schemas:
...
    Content:
      type: object
      required:
        - id
        - userId
        - walletAddress
        - contentType
        - status
        - title
        - description
        - tags
        - game
        - platform
        - category
        - files
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          format: uuid
          description: Content creator ID
        walletAddress:
          type: string
          pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
          description: Creator's Solana wallet address
          example: "7XJiwLDrjzxDYdZipnJXzpr1iDTmK55XixSFAa7JgNEL"
        contentType:
          $ref: '#/components/schemas/ContentType'
        status:
          $ref: '#/components/schemas/ContentStatus'
        title:
          type: string
          minLength: 3
          maxLength: 100
          example: "Epic Fortnite Victory Royale - Solo Clutch"
        description:
          type: string
          minLength: 10
          maxLength: 1000
          example: "Incredible last-minute clutch in Fortnite Battle Royale. Watch as I take down the final squad with only 10 HP!"
        tags:
          type: array
          items:
            type: string
            maxLength: 25
          minItems: 1
          maxItems: 15
          example: ["fortnite", "victory-royale", "clutch", "solo"]
        game:
          type: string
          minLength: 2
          maxLength: 100
          example: "Fortnite"
        platform:
          $ref: '#/components/schemas/Platform'
        category:
          $ref: '#/components/schemas/Category'
        gameMode:
          type: string
          maxLength: 50
          example: "Solo Battle Royale"
        difficulty:
          type: string
          enum: [easy, medium, hard, expert, legendary]
        files:
          type: array
          items:
            $ref: '#/components/schemas/ContentFile'
          minItems: 1
          maxItems: 5
        thumbnailUrl:
          type: string
          format: uri
          example: "https://cdn.mlg.clan/thumbnails/content-123.jpg"
        previewUrl:
          type: string
          format: uri
          example: "https://cdn.mlg.clan/previews/content-123.gif"
        views:
          type: integer
          minimum: 0
          example: 1250
        likes:
          type: integer
          minimum: 0
          example: 89
        dislikes:
          type: integer
          minimum: 0
          example: 3
        shares:
          type: integer
          minimum: 0
          example: 12
        comments:
          type: integer
          minimum: 0
          example: 15
        mlgVotes:
          $ref: '#/components/schemas/MLGVoteStats'
        qualityScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 87.5
        viralScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 65.2
        trendingScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 92.3
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T12:30:00Z"
        publishedAt:
          type: string
          format: date-time
          example: "2024-01-01T12:15:00Z"
        moderationStatus:
          type: string
          enum: [approved, pending, flagged, rejected]
          default: pending
        reportCount:
          type: integer
          minimum: 0
          example: 0
    ContentFile:
      type: object
      required:
        - id
        - filename
        - originalName
        - mimeType
        - size
        - url
        - processingStatus
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          example: "epic-victory-royale.mp4"
        originalName:
          type: string
          example: "My Epic Game Clip.mp4"
        mimeType:
          type: string
          example: "video/mp4"
        size:
          type: integer
          minimum: 1
          example: 15728640
        url:
          type: string
          format: uri
          example: "https://cdn.mlg.clan/videos/epic-victory-royale.mp4"
        cdnUrl:
          type: string
          format: uri
          example: "https://cdn.mlg.clan/videos/epic-victory-royale.mp4"
        thumbnailUrl:
          type: string
          format: uri
          example: "https://thumbnails.mlg.clan/epic-victory-royale.jpg"
        duration:
          type: number
          format: float
          minimum: 0
          example: 45.5
          description: Duration in seconds (for video/audio files)
        dimensions:
          type: object
          properties:
            width:
              type: integer
              minimum: 1
              example: 1920
            height:
              type: integer
              minimum: 1
              example: 1080
        processingStatus:
          type: string
          enum: [pending, processing, complete, failed]
          example: complete
        processedFormats:
          type: array
          items:
            type: string
          example: ["360p", "720p", "1080p"]
        createdAt:
          type: string
          format: date-time
    MLGVoteStats:
      type: object
      required:
        - upvotes
        - downvotes
        - superVotes
        - totalTokensBurned
      properties:
        upvotes:
          type: integer
          minimum: 0
          example: 42
        downvotes:
          type: integer
          minimum: 0
          example: 2
        superVotes:
          type: integer
          minimum: 0
          example: 5
        totalTokensBurned:
          type: number
          format: float
          minimum: 0
          example: 67.0
...
    ContentCreateInput:
      type: object
      required:
        - title
        - description
        - contentType
        - game
        - platform
        - category
        - tags
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          minLength: 10
          maxLength: 1000
        contentType:
          $ref: '#/components/schemas/ContentType'
        game:
          type: string
          minLength: 2
          maxLength: 100
        platform:
          $ref: '#/components/schemas/Platform'
        category:
          $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            type: string
            maxLength: 25
          minItems: 1
          maxItems: 15
        gameMode:
          type: string
          maxLength: 50
        difficulty:
          type: string
          enum: [easy, medium, hard, expert, legendary]
        isPublic:
          type: boolean
          default: true
        allowComments:
          type: boolean
          default: true
        allowVoting:
          type: boolean
          default: true
...
    ContentType:
      type: string
      enum:
        - video_clip
        - screenshot
        - guide
        - review
        - stream
        - tournament_content
      example: video_clip
    ContentStatus:
      type: string
      enum:
        - draft
        - pending
        - published
        - flagged
        - removed
      example: published
...
    Platform:
      type: string
      enum:
        - pc
        - xbox
        - playstation
        - mobile
        - nintendo
        - steam-deck
        - other
      example: pc
    Category:
      type: string
      enum:
        - highlights
        - gameplay
        - tutorials
        - funny
        - competitive
        - speedrun
        - review
        - guide
      example: highlights

    # Response Schemas
    SuccessResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
            version:
              type: string
              example: "1.0.0"
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - meta
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "CONTENT_001"
        message:
          type: string
          example: "Content not found"
        details:
          type: object
          description: Additional error details
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
            version:
              type: string
              example: "1.0.0"
    ValidationError:
      type: object
      required:
        - field
        - message
        - code
      properties:
        field:
          type: string
          example: "title"
        message:
          type: string
          example: "Title is required"
        code:
          type: string
          example: "VALIDATION_001"
        details:
          type: object
    PaginationResponse:
      type: object
      required:
        - total
        - limit
        - offset
        - hasNext
        - hasPrev
      properties:
        total:
          type: integer
          minimum: 0
          example: 150
        limit:
          type: integer
          minimum: 1
          example: 20
        offset:
          type: integer
          minimum: 0
          example: 40
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: true
        nextCursor:
          type: string
          example: "eyJpZCI6IjEyMyIsInRzIjoxNjQwOTk1MjAwfQ"
        prevCursor:
          type: string
          example: "eyJpZCI6IjEwMCIsInRzIjoxNjQwOTk1MTAwfQ"
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "VALIDATION_002"
            message: "Invalid field format"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "AUTH_001"
            message: "Invalid or expired JWT token"
...
    PayloadTooLarge:
      description: Payload too large
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "FILE_001"
            message: "File too large"
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "RATE_001"
            message: "Too many requests"
            details:
              limit: 100
              remaining: 0
              resetTime: "2024-01-01T13:00:00Z"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  details:
                    type: object
                    properties:
                      errors:
                        type: array
                        items:
                          $ref: '#/components/schemas/ValidationError'
                      warnings:
                        type: array
                        items:
                          $ref: '#/components/schemas/ValidationError'
          example:
            success: false
            error: "validation_failed"
            message: "Content validation failed"
            details:
              errors:
                - field: "title"
                  message: "Title is required"
                  code: "VALIDATION_001"
                - field: "files"
                  message: "At least one file is required"
                  code: "VALIDATION_001"
...